// web/api/src/contracts/facturas.ts
export type Factura = {
  NUM_FACT: string;
  SERIALTIPO: string;
  Tipo_Orden: string;
  CODIGO: string | null;
  FECHA: Date | null;
  FECHA_veN: Date | null;
  HORA: string | null;
  NOMBRE: string | null;
  MONTO_GRA: number | null;
  IVA: number | null;
  MONTO_EXE: number | null;
  TOTAL: number | null;
  PAGO: string | null;
  ORDEN: string | null;
  CHEQUE: number | null;
  BANCO_CHEQUE: string | null;
  NOTA: string | null;
  ANULADA: boolean;
  OBSERV: string | null;
  CARGO: number | null;
  PEDIDO: string | null;
  FECHA_ORDEN: Date | null;
  MONTO_ORDEN: number | null;
  LOCACION: string | null;
  INSPECTOR: string | null;
  REPORTE: string | null;
  FECHA_REPORTE: Date | null;
  ALICUOTA: number | null;
  CANCELADA: string | null;
  Relacionada: boolean;
  Monto_Efect: number | null;
  Monto_Cheque: number | null;
  Banco_Tarjeta: string | null;
  Tarjeta: string | null;
  Monto_Tarjeta: number | null;
  COD_USUARIO: string | null;
  Placas: string | null;
  Kilometros: number | null;
  Fecha_llamado: Date | null;
  Vino_30: boolean;
  Vino_60: boolean;
  Vino_90: boolean;
  Saldo: number | null;
  Abono: number | null;
  Moneda: string | null;
  Tasacambio: number | null;
  Recibido: string | null;
  Entregado: string | null;
  Cta: string | null;
  Vendedor: string | null;
  Peaje: number | null;
  tasa_dolar: number | null;
  Num_Control: string | null;
  RetencionIva: number | null;
  Terminos: string | null;
  Despachar: string | null;
  FOB: string | null;
  departamento: string | null;
  RIF: string | null;
  Nro_Retencion: string | null;
  Monto_Retencion: number | null;
  Fecha_Retencion: Date | null;
  Cancelado: number | null;
  Vuelto: number | null;
  TIPO_RET: string | null;
  MONTO_GRABS: number | null;
  TOTALPAGO: number | null;
  FECHAANULADA: Date | null;
};

export type DetalleFactura = {
  Id: number;
  NUM_FACT: string | null;
  SERIALTIPO: string | null;
  COD_SERV: string | null;
  DESCRIPCION: string | null;
  FECHA: Date | null;
  CANTIDAD: number | null;
  PRECIO: number | null;
  TOTAL: number | null;
  ANULADA: boolean | null;
  Co_Usuario: string | null;
  HORA: string | null;
  NOTA: number | null;
  unidad: string | null;
  Alicuota: number | null;
  PRECIO_DESCUENTO: number | null;
  DESCUENTO: number | null;
  Relacionada: boolean;
  RENGLON: number | null;
  SERVICIO: number | null;
  Vendedor: string | null;
  Autoriza: string | null;
  FECHAANULADA: Date | null;
  Cod_Alterno: string | null;
  PlacaSerial: string | null;
  Rif_Propietario: string | null;
  Nombre_Propietario: string | null;
  Marca_Vehiculo: string | null;
  Modelo: string | null;
  Fabricacion: string | null;
  Almacen: string | null;
  Lote: string | null;
  CostoLote: number | null;
  fechalote: Date | null;
  facturalote: string | null;
  Cantidadlote: number | null;
  Comision: number | null;
  TasaCambio: number | null;
};